requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
betterproto==2.0.0b7
certifi==2026.2.25
colorama==0.4.6
distro==1.9.0
edge-tts==7.2.7
ffmpy==1.0.0
frozenlist==1.8.0
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
mashumaro==3.20
multidict==6.7.1
openai==2.26.0
playsound==1.2.2
propcache==0.4.1
protobuf==7.34.0
protobuf3-to-dict==0.1.5
pydantic==2.12.5
pydantic_core==2.41.5
pyee==11.1.0
pytchat==0.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-socks==2.8.1
six==1.17.0
sniffio==1.3.1
tabulate==0.10.0
TikTokLive==6.1.0
tqdm==4.67.3
twitchio==3.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
websockets==16.0
websockets_proxy==0.1.3
yarl==1.23.0
youtube-search-python==1.6.6
